+++ src/main/webapp/admin/groupnoacl.jsp
@@ -65,17 +65,20 @@
     	}
     	if(providerNos != null) {    		    	
 	    	for(String providerNoTemp : providerNos) {
-	    		MyGroupAccessRestriction mgra = myGroupAccessRestrictionDao.findByGroupNoAndProvider(chosen_group,providerNoTemp);
-	    		if(mgra != null) {
-	    			myGroupAccessRestrictionDao.remove(mgra.getId());
-	    		}
-	    		
-	   			mgra = new MyGroupAccessRestriction();
-	   			mgra.setMyGroupNo(chosen_group);
-	   			mgra.setProviderNo(providerNoTemp);
-	   			mgra.setLastUpdateUser(providerNoTemp);
-	       		myGroupAccessRestrictionDao.persist(mgra);
-	    	}
+				if (chosen_group.length() <= 50 &&
+					providerNoTemp.length() <= 6) {
+					MyGroupAccessRestriction mgra = myGroupAccessRestrictionDao.findByGroupNoAndProvider(chosen_group, providerNoTemp);
+					if (mgra != null) {
+						myGroupAccessRestrictionDao.remove(mgra.getId());
+					}
+
+					mgra = new MyGroupAccessRestriction();
+					mgra.setMyGroupNo(chosen_group);
+					mgra.setProviderNo(providerNoTemp);
+					mgra.setLastUpdateUser(providerNoTemp);
+					myGroupAccessRestrictionDao.persist(mgra);
+				}
+			}
     	}
     }
